requirements.txt
langchain
langchain-community
faiss-cpu
flask
flask-restful
beautifulsoup4
html5lib
openai
python-dotenv
requests